casetest: split DAG test
