+++ planner/core/casetest/dag/BUILD.bazel
@@ -1,36 +1,30 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
-    name = "physicalplantest_test",
+    name = "dag_test",
     timeout = "short",
     srcs = [
+        "dag_test.go",
         "main_test.go",
-        "physical_plan_test.go",
     ],
     data = glob(["testdata/**"]),
     flaky = True,
-    race = "on",
-    shard_count = 50,
+    shard_count = 10,
     deps = [
-        "//config",
         "//domain",
-        "//executor",
         "//infoschema",
         "//parser",
         "//parser/ast",
         "//parser/model",
         "//planner",
         "//planner/core",
-        "//planner/core/internal",
-        "//sessionctx/stmtctx",
-        "//sessionctx/variable",
+        "//session",
+        "//sessiontxn",
         "//testkit",
-        "//testkit/external",
         "//testkit/testdata",
         "//testkit/testmain",
         "//testkit/testsetup",
         "//util/hint",
-        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
