+++ planner/core/casetest/physicalplantest/physical_plan_test.go
@@ -31,10 +31,8 @@ import (
 	"github.com/pingcap/tidb/planner"
 	"github.com/pingcap/tidb/planner/core"
 	"github.com/pingcap/tidb/planner/core/internal"
-	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx/stmtctx"
 	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/sessiontxn"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/testkit/external"
 	"github.com/pingcap/tidb/testkit/testdata"
@@ -101,286 +99,6 @@ func doTestPushdownDistinct(t *testing.T, vars, input []string, output []struct 
 		})
 		tk.MustQuery("explain format = 'brief' " + ts).Check(testkit.Rows(output[i].Plan...))
 		tk.MustQuery(ts).Sort().Check(testkit.Rows(output[i].Result...))
-	}
-}
-
-func TestDAGPlanBuilderSimpleCase(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("set tidb_opt_limit_push_down_threshold=0")
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("case: %v, sql: %s", i, tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-		require.NoError(t, sessiontxn.NewTxn(context.Background(), tk.Session()))
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), comment)
-	}
-}
-
-func TestDAGPlanBuilderJoin(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	sessionVars := tk.Session().GetSessionVars()
-	sessionVars.ExecutorConcurrency = 4
-	sessionVars.SetDistSQLScanConcurrency(15)
-	sessionVars.SetHashJoinConcurrency(5)
-
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("case:%v sql:%s", i, tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), comment)
-	}
-}
-
-func TestDAGPlanBuilderSubquery(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("set sql_mode='STRICT_TRANS_TABLES'") // disable only full group by
-	sessionVars := tk.Session().GetSessionVars()
-	sessionVars.SetHashAggFinalConcurrency(1)
-	sessionVars.SetHashAggPartialConcurrency(1)
-	sessionVars.SetHashJoinConcurrency(5)
-	sessionVars.SetDistSQLScanConcurrency(15)
-	sessionVars.ExecutorConcurrency = 4
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("input: %s", tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), fmt.Sprintf("input: %s", tt))
-	}
-}
-
-func TestDAGPlanTopN(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("case:%v sql:%s", i, tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), comment)
-	}
-}
-
-func TestDAGPlanBuilderBasePhysicalPlan(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	se, err := session.CreateSession4Test(store)
-	require.NoError(t, err)
-	_, err = se.Execute(context.Background(), "use test")
-	require.NoError(t, err)
-
-	var input []string
-	var output []struct {
-		SQL   string
-		Best  string
-		Hints string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("input: %s", tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		err = core.Preprocess(context.Background(), se, stmt, core.WithPreprocessorReturn(&core.PreprocessorReturn{InfoSchema: is}))
-		require.NoError(t, err)
-		p, _, err := planner.Optimize(context.TODO(), se, stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-			output[i].Hints = hint.RestoreOptimizerHints(core.GenHintsFromPhysicalPlan(p))
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), fmt.Sprintf("input: %s", tt))
-		hints := core.GenHintsFromPhysicalPlan(p)
-
-		// test the new genHints code
-		flat := core.FlattenPhysicalPlan(p, false)
-		newHints := core.GenHintsFromFlatPlan(flat)
-		assertSameHints(t, hints, newHints)
-
-		require.Equal(t, output[i].Hints, hint.RestoreOptimizerHints(hints), fmt.Sprintf("input: %s", tt))
-	}
-}
-
-func TestDAGPlanBuilderUnion(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("case:%v sql:%s", i, tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), comment)
-	}
-}
-
-func TestDAGPlanBuilderUnionScan(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, c int)")
-
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-
-	p := parser.New()
-	for i, tt := range input {
-		tk.MustExec("begin;")
-		tk.MustExec("insert into t values(2, 2, 2);")
-
-		comment := fmt.Sprintf("input: %s", tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-		dom := domain.GetDomain(tk.Session())
-		require.NoError(t, dom.Reload())
-		plan, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, dom.InfoSchema())
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(plan)
-		})
-		require.Equal(t, output[i].Best, core.ToString(plan), fmt.Sprintf("input: %s", tt))
-		tk.MustExec("rollback;")
-	}
-}
-
-func TestDAGPlanBuilderAgg(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("set sql_mode='STRICT_TRANS_TABLES'") // disable only full group by
-	sessionVars := tk.Session().GetSessionVars()
-	sessionVars.SetHashAggFinalConcurrency(1)
-	sessionVars.SetHashAggPartialConcurrency(1)
-	sessionVars.SetDistSQLScanConcurrency(15)
-	sessionVars.ExecutorConcurrency = 4
-
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-	for i, tt := range input {
-		comment := fmt.Sprintf("input: %s", tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), fmt.Sprintf("input: %s", tt))
 	}
 }
 
@@ -1847,66 +1565,6 @@ func TestIndexJoinHint(t *testing.T) {
 		})
 		require.Equal(t, output[i].Plan, core.ToString(p), comment)
 	}
-}
-
-func doTestDAGPlanBuilderWindow(t *testing.T, vars, input []string, output []struct {
-	SQL  string
-	Best string
-}) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	for _, v := range vars {
-		tk.MustExec(v)
-	}
-
-	p := parser.New()
-	is := infoschema.MockInfoSchema([]*model.TableInfo{core.MockSignedTable(), core.MockUnsignedTable()})
-
-	for i, tt := range input {
-		comment := fmt.Sprintf("case:%v sql:%s", i, tt)
-		stmt, err := p.ParseOneStmt(tt, "", "")
-		require.NoError(t, err, comment)
-
-		err = sessiontxn.NewTxn(context.Background(), tk.Session())
-		require.NoError(t, err)
-		p, _, err := planner.Optimize(context.TODO(), tk.Session(), stmt, is)
-		require.NoError(t, err)
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Best = core.ToString(p)
-		})
-		require.Equal(t, output[i].Best, core.ToString(p), comment)
-	}
-}
-
-func TestDAGPlanBuilderWindow(t *testing.T) {
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	vars := []string{
-		"set @@session.tidb_window_concurrency = 1",
-	}
-	doTestDAGPlanBuilderWindow(t, vars, input, output)
-}
-
-func TestDAGPlanBuilderWindowParallel(t *testing.T) {
-	var input []string
-	var output []struct {
-		SQL  string
-		Best string
-	}
-	planSuiteData := GetPlanSuiteData()
-	planSuiteData.LoadTestCases(t, &input, &output)
-	vars := []string{
-		"set @@session.tidb_window_concurrency = 4",
-	}
-	doTestDAGPlanBuilderWindow(t, vars, input, output)
 }
 
 func TestNominalSort(t *testing.T) {
